Add MonadThreeCol layout
Introduce a new layout: MonadThreeCol. It's based on XMonad's
ThreeColumns layout and it's useful for people using ultra wide
displays. It behaves similarly to MonadTall with two key differences:
the main window can be placed at the center of the screen and the
secondary windows are divided into two columns.